chore(refactoring): More Refactoring of Scheduler Component
