+++ src/components/Scheuler2/scheduler.controller.js
@@ -2,22 +2,46 @@ import { Loader } from "redux/dispatcher/Loader"
 import { useSchedulerModel } from "./scheduler.model"
 import { useState } from "react"
 import { getDataArray, getEventListing } from "helpers/conversionFunctions/conversion"
-// import { getDataArray } from "helpers/conversionFunctions/conversion"
-
+import { useMediaQuery, useTheme } from "@mui/material"
+import dayjs from "dayjs"
+import { DATE_FORMAT } from "BigScheduler"
 export const useSchedulerController = () => {
+  const theme = useTheme()
   const [departments, setDepartments] = useState([])
   const [teamMembers, setTeamMemebers] = useState([])
   // const [teamSchedules, setTeamSchedules] = useState([])
   const [projects, setProjects] = useState([])
   const [clients, setClients] = useState([])
   const [reload, setReload] = useState(false)
+  const [removeResource, setRemoveResource] = useState(false)
+  const [rerender, triggerRerender] = useState(1)
   //eslint-disable-next-line no-unused-vars
   const [teamMemberArray, setTeamMemberArray] = useState(false)
+  const [schedulerData, setSchedulerData] = useState(null)
+  const [triger, setRetrigger] = useState(false)
+  const [popupChild, setPopupChild] = useState("")
+  const [openPopUp, setOpenPopup] = useState(true)
+  const [view, setView] = useState(1)
+  const [id, setId] = useState("")
+  const [resoureMap, setResourceMap] = useState(new Map())
+  const [eventsMap, setEventsMap] = useState(new Map())
+  const [selectedObject, setSelectedObject] = useState(null)
+  const [popupStyles, setPopUpStyles] = useState({})
+  const [isAddeventPopover, setIsAddeventPopover] = useState(false)
+  const [resourceEvent, setResourceEvent] = useState(null)
+  const isMobile = useMediaQuery(theme.breakpoints.down("sm"))
+  const isTablet = useMediaQuery(theme.breakpoints.down("md"))
+  const [counter, setCounter] = useState(1)
+  const [fetchEvents, setFetchEvents] = useState(false)
+  const [rerenderData, setRerenderData] = useState(false)
+  const [startDate, setStartDate] = useState(new dayjs(new Date()).format(DATE_FORMAT))
+  const [fetcher, setFetcher] = useState(false)
+  const [search, setSearch] = useState("")
+
   const model = useSchedulerModel()
   const fetchDepartments = async () => {
     const data = await model.fetchDepartments()
     if (data?.length > 0) {
-      // if (response?.success && response?.code === 200) {
       const requiredData = data.map((department) => {
         return {
           label: department?.name,
@@ -38,6 +62,7 @@ export const useSchedulerController = () => {
       requiredArray = []
     }
     setTeamMemebers(requiredArray)
+    setRemoveResource(false)
     setReload((prev) => !prev)
   }
   const fetchSchedules = async (params) => {
@@ -47,7 +72,6 @@ export const useSchedulerController = () => {
     if (responseData?.data?.length > 0) {
       return { data: getEventListing(responseData?.data), success: responseData?.success }
     }
-    // setTeamSchedules(requiredArray)
   }
   const updateSchedules = async (params, body) => {
     const responseData = await model.updateSchedule(params, body)
@@ -145,6 +169,69 @@ export const useSchedulerController = () => {
       bgColor: `${parentObject?.color}`
     }
   }
+  const handlePopUpClose = () => {
+    setOpenPopup(false)
+    setPopupChild("")
+    setIsAddeventPopover(false)
+    setPopUpStyles(null)
+    setResourceEvent(null)
+    if (removeResource) {
+      setFetcher((prev) => !prev)
+    }
+  }
+  const handlePopover = (check, newStyles) => {
+    if (check) {
+      setIsAddeventPopover(true)
+      setPopUpStyles(newStyles)
+    } else {
+      setOpenPopup(true)
+    }
+  }
+  const keepDataOpen = (openArrays, sd) => {
+    openArrays.forEach((arrayItem) => {
+      toggleExpandFunc(sd, arrayItem?.slotId, true)
+    })
+  }
+  const toggleExpandFunc = (schedulerData, slotId, value) => {
+    schedulerData.toggleExpandStatus(slotId, value)
+    triggerRerender(rerender + 1)
+    setView(view + 1)
+  }
+  const getBgColor = (id, resourceId) => {
+    const responseMap = resoureMap.get(id)
+    const projectArray = responseMap.map((item) => item?.projects)
+    const flatArray = projectArray.flat()
+    const filteredArray = flatArray.filter((item) => item?.id === resourceId)
+    return filteredArray[0]?.color
+  }
+  const handlePopUp = (key) => {
+    switch (key) {
+      case "cal":
+        setPopupChild("calenderFeed")
+        setOpenPopup(true)
+        return
+      case "del":
+        setPopupChild("deleteResource")
+        setOpenPopup(true)
+        return
+      case "edit":
+        setPopupChild("editResource")
+        setOpenPopup(true)
+        return
+      case "arc":
+        setPopupChild("archiveResource")
+        setOpenPopup(true)
+        return
+      case "add":
+        setPopupChild("projectForm")
+        setOpenPopup(true)
+        return
+      case "editEvent":
+        setPopupChild("editEvent")
+        setOpenPopup(true)
+        return
+    }
+  }
   return {
     fetchDepartments,
     departments,
@@ -166,6 +253,54 @@ export const useSchedulerController = () => {
     reloader,
     getChildObjectArray,
     getFreshId,
-    newEventObject
+    newEventObject,
+    removeResource,
+    setRemoveResource,
+    rerender,
+    triggerRerender,
+    schedulerData,
+    setSchedulerData,
+    popupChild,
+    setPopupChild,
+    openPopUp,
+    setOpenPopup,
+    view,
+    setView,
+    id,
+    setId,
+    resoureMap,
+    setResourceMap,
+    eventsMap,
+    setEventsMap,
+    selectedObject,
+    setSelectedObject,
+    popupStyles,
+    setPopUpStyles,
+    isAddeventPopover,
+    setIsAddeventPopover,
+    resourceEvent,
+    setResourceEvent,
+    isMobile,
+    isTablet,
+    counter,
+    setCounter,
+    fetchEvents,
+    setFetchEvents,
+    rerenderData,
+    setRerenderData,
+    startDate,
+    setStartDate,
+    fetcher,
+    setFetcher,
+    search,
+    setSearch,
+    triger,
+    setRetrigger,
+    handlePopUpClose,
+    handlePopover,
+    keepDataOpen,
+    toggleExpandFunc,
+    getBgColor,
+    handlePopUp
   }
 }
